requirements.txt
discord.py==1.6.0
python-dotenv==0.15.0
discord_lumberjack==1.0.5